Move filesystem related methods and const from Boost_Cache_Utils.php to Filesystem_Utils.php
