+++ projects/plugins/boost/app/modules/optimizations/page-cache/pre-wordpress/Filesystem_Utils.php
@@ -3,6 +3,84 @@
 namespace Automattic\Jetpack_Boost\Modules\Optimizations\Page_Cache\Pre_WordPress;
 
 class Filesystem_Utils {
+
+	const DELETE_ALL   = 'delete-all'; // delete all files and directories in a given directory, recursively.
+	const DELETE_FILE  = 'delete-single'; // delete a single file or recursively delete a single directory in a given directory.
+	const DELETE_FILES = 'delete-files'; // delete all files in a given directory.
+
+	/**
+	 * Recursively delete a directory.
+	 * @param string $path - The directory to delete.
+	 * @param bool   $type - The type of delete. DELETE_FILES to delete all files in the given directory. DELETE_ALL to delete everything in the given directory, recursively.
+	 * @return bool|WP_Error
+	 */
+	public static function delete_directory( $path, $type ) {
+		Logger::debug( "delete directory: $path $type" );
+		$path = realpath( $path );
+		if ( ! $path ) {
+			// translators: %s is the directory that does not exist.
+			return new \WP_Error( 'directory-missing', sprintf( __( 'Directory does not exist: %s', 'jetpack-boost' ), $path ) ); // realpath returns false if a file does not exist.
+		}
+
+		// make sure that $dir is a directory inside WP_CONTENT . '/boost-cache/';
+		if ( self::is_boost_cache_directory( $path ) === false ) {
+			// translators: %s is the directory that is invalid.
+			return new \WP_Error( 'invalid-directory', sprintf( __( 'Invalid directory %s', 'jetpack-boost' ), $path ) );
+		}
+
+		if ( ! is_dir( $path ) ) {
+			return new \WP_Error( 'not-a-directory', __( 'Not a directory', 'jetpack-boost' ) );
+		}
+
+		switch ( $type ) {
+			case self::DELETE_ALL: // delete all files and directories in the given directory.
+				$iterator = new \RecursiveIteratorIterator( new \RecursiveDirectoryIterator( $path, \RecursiveDirectoryIterator::SKIP_DOTS ) );
+				foreach ( $iterator as $file ) {
+					if ( $file->isDir() ) {
+						Logger::debug( 'rmdir: ' . $file->getPathname() );
+						@rmdir( $file->getPathname() ); // phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_rmdir, WordPress.PHP.NoSilencedErrors.Discouraged
+					} else {
+						Logger::debug( 'unlink: ' . $file->getPathname() );
+						@unlink( $file->getPathname() ); // phpcs:ignore WordPress.WP.AlternativeFunctions.unlink_unlink, WordPress.PHP.NoSilencedErrors.Discouraged
+					}
+				}
+				@rmdir( $path ); // phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_rmdir, WordPress.PHP.NoSilencedErrors.Discouraged,
+				break;
+			case self::DELETE_FILES: // delete all files in the given directory.
+				$files = array_diff( scandir( $path ), array( '.', '..' ) );
+				foreach ( $files as $file ) {
+					$file = $path . '/' . $file;
+					if ( is_file( $file ) ) {
+						Logger::debug( "unlink: $file" );
+						@unlink( $file ); // phpcs:ignore WordPress.PHP.NoSilencedErrors.Discouraged, WordPress.WP.AlternativeFunctions.unlink_unlink
+					}
+				}
+				break;
+		}
+		return true;
+	}
+
+	/**
+	 * Returns true if the given directory is inside the boost-cache directory.
+	 * @param string $dir - The directory to check.
+	 * @return bool
+	 */
+	public static function is_boost_cache_directory( $dir ) {
+		$dir = self::sanitize_file_path( $dir );
+		return strpos( $dir, WP_CONTENT_DIR . '/boost-cache' ) !== false;
+	}
+
+	/**
+	 * Given a request_uri and its parameters, return the filename to use for this cached data. Does not include the file path.
+	 *
+	 * @param array  $parameters  - An associative array of all the things that make this request special/different. Includes GET parameters and COOKIEs normally.
+	 */
+	public static function get_request_filename( $parameters ) {
+
+		$key_components = apply_filters( 'boost_cache_key_components', $parameters );
+
+		return md5( json_encode( $key_components ) ) . '.html'; // phpcs:ignore WordPress.WP.AlternativeFunctions.json_encode_json_encode
+	}
 	/**
 	 * Recursively garbage collect a directory.
 	 *
